+++ HISTORY.rst
@@ -1,7 +1,7 @@
 History
 =======
 
-0.3.1 (unreleased)
+0.4.0 (unreleased)
 ------------------
 
 - bugfix: if a row uses an ``I`` action, everything there is ignored
@@ -11,6 +11,9 @@ History
   Closes #12
 - Human readable error message when the sheet is not found.
   Closes #11
+- Added ``-p`` parameter, to just act on specific project(s)
+- Added ``--action``, and implemented ``report`` action (credits to @gcammarota)
+- More detailed logging when deleting events with action ``D``
 
 0.3.0 (2022-05-02)
 ------------------